feat: comment 반영 - 파일 명명법 수정 및 mock-server 데이터로 수정
